# Improve elementwise performance.